2/3 SI Robot Testing
What we did: We checked each subsystem and created  methods to control the intake and arm via percent output to check that the code works. We also created commands to check functionality, and will transition to check if motion magic works as well.
TODO: Look at the steps provided by Aalind on sheets, and make sure EVERYTHING WORKS